Exposing ChaCha20Poly1305 for windows browser (#23)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/72649045549333/task/1210463292173324?focus=true
Tech Design URL:
cc: @lezzi 

Since the .NET ChaCha20Poly1305 is not supported on most of Windows 10
versions, we need to expose the ChaCha20Poly1305 implementation for
Windows Browser and for that we utilize the sync crypto DLL which
already use libsodium.

An extra, I added the first testing project and also bumped to .net 8,
which has been already changed in windows-browser.

Steps to test this PR
---------------------
1. Run the unit tests for WindowsBrowser.Sync.Crypto.Managed.Tests